[Fix] 생성 친구 관련 API, 메서드 삭제 #47
