+++ favor/src/main/java/com/favor/favor/friend/Friend.java
@@ -45,20 +45,6 @@ public class Friend extends TimeStamped {
     @OneToMany(mappedBy = "friend", orphanRemoval = true)
     private List<Reminder> reminderList = new ArrayList<>();
 
-
-    @Builder.Default
-    @ElementCollection
-    private List<Integer> favorList = new ArrayList<>();
-    public void setFavorList(List<Favor> favorList){
-        ArrayList<Integer> favorTypeList = new ArrayList<>();
-        for(Favor favor : favorList){
-            favorTypeList.add(favor.getType());
-        }
-        this.favorList = favorTypeList;
-    }
-
-    private Boolean isUser;
-
     private Long friendUserNo;
 
     @Builder.Default
