+++ favor/src/main/java/com/favor/favor/friend/FriendResponseDto.java
@@ -1,6 +1,5 @@
 package com.favor.favor.friend;
 
-import com.favor.favor.anniversary.AnniversaryResponseDto;
 import com.favor.favor.common.enums.Favor;
 import com.favor.favor.gift.GiftResponseDto;
 import com.favor.favor.reminder.ReminderResponseDto;
@@ -12,8 +11,6 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 public class FriendResponseDto {
-
-    private Boolean isUser;
 
     private Long friendNo;
     private String friendName;
@@ -32,7 +29,6 @@ public class FriendResponseDto {
 
     @Builder
     public FriendResponseDto(Friend friend){
-        this.isUser = friend.getIsUser();
         this.friendNo = friend.getFriendNo();
         this.friendName = friend.getFriendName();
         this.friendMemo = friend.getFriendMemo();
@@ -51,7 +47,6 @@ public class FriendResponseDto {
                              List<Favor> favorList,
                              List<Long> anniversaryNoList
     ){
-        this.isUser = friend.getIsUser();
         this.friendNo = friend.getFriendNo();
         this.friendName = friend.getFriendName();
         this.friendMemo = friend.getFriendMemo();
