--- favor/src/main/java/com/favor/favor/friend/account/FriendUserRequestDto.java
@@ -1,28 +1,0 @@
-package com.favor.favor.friend.account;
-
-import com.favor.favor.friend.Friend;
-import com.favor.favor.user.User;
-import io.swagger.annotations.ApiModelProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.transaction.Transactional;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class FriendUserRequestDto {
-    @ApiModelProperty(position = 1, required = true, dataType = "Long", value = "회원친구번호", example = "1")
-    private Long friendUserNo;
-    @Transactional
-    public Friend toEntity(User user, User userFriend){
-        return Friend.builder()
-                .isUser(true)
-                .friendUserNo(friendUserNo)
-                .friendName(userFriend.getName())
-                .friendMemo("")
-                .user(user)
-                .build();
-    }
-}
